Replace AccountGeneralPreferences entity with git backend (phase 1)
Replace the usage of AccountGeneralPreferences with a different class
in extension API package.

To support live migration, the upgrade is done in two steps:

* phase 1 (this change):
  o Always write to both git and db
  o Introduce new configuration option to indicate whether to read from
    git or db, initially set to read from db
  o First binary update: some servers are reading/writing just the db;
    some servers are additionally writing to git
  o After first update: all servers are reading from the db, writing to
    both
  o Batch copy data from db to git (only related to non open source
    Gerrit version)
  o Update all servers to read from git. During the update, some will
    still be reading from the db; that's ok, because everybody is
    writing to git
  o OutgoingEmail class is retrieving copy_self_on_email preference
    always form the database (by getting AccountGlobalPreferences from
    the account cache). This is OK, because all servers are writing to
    the database
* phase 2 (next change):
  o Bump database version, migrate the data from db to git, delete the
    entity (and the flag) from the code and update the servers.

+++ gerrit-gwtui/src/main/java/com/google/gerrit/client/account/AccountMessages.java
@@ -18,7 +18,7 @@ import com.google.gwt.i18n.client.Messages;
 
 public interface AccountMessages extends Messages {
   String lines(short cnt);
-  String rowsPerPage(short cnt);
+  String rowsPerPage(int cnt);
   String changeScreenServerDefault(String d);
   String enterIAGREE(String iagree);
 }
